requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
openai-whisper
easyocr
numpy<2.0.0
python-dotenv
ddgs
duckduckgo-search
wikipedia
groq
torch
torchvision
transformers
sentence-transformers
opencv-python-headless
Pillow
